Разделили на модули текущее и добавили будущие модули
