+++ js/data-transfer.js
@@ -1,0 +1,2 @@
+//На будущее
+//Отправка данных на сервер
