+++ js/edit-image.js
@@ -1,0 +1,2 @@
+//На будущее
+//Редактирование изображения и ограничения, накладываемые на поля
