+++ js/filter-image.js
@@ -1,0 +1,2 @@
+//На будущее
+//Фильтрация изображений от других пользователей
