+++ js/upload-new-photo.js
@@ -1,0 +1,2 @@
+//На будущее
+//Загрузка нового изображения на сайт и заполнение информации о нём
