+++ js/viewer-image.js
@@ -1,0 +1,2 @@
+//На будущее
+//Просмотр загруженных изображений
